feat: add ChannelTable widget
