+++ src/pymmcore_widgets/_general_mda_widgets.py
@@ -22,122 +22,6 @@ from qtpy.QtWidgets import (
 from superqt.fonticon import icon
 
 
-class _MDAChannelTable(QGroupBox):
-
-    valueUpdated = Signal()
-
-    def __init__(self, *, parent: QWidget | None = None) -> None:
-        super().__init__(parent=parent)
-
-        self._mmc = CMMCorePlus.instance()
-
-        self._create_ch_gui()
-
-    def _create_ch_gui(self) -> None:
-
-        self.setTitle("Channels")
-
-        group_layout = QGridLayout()
-        group_layout.setSpacing(15)
-        group_layout.setContentsMargins(10, 10, 10, 10)
-        self.setLayout(group_layout)
-
-        # channel table
-        self.channel_tableWidget = QTableWidget()
-        hdr = self.channel_tableWidget.horizontalHeader()
-        hdr.setSectionResizeMode(hdr.ResizeMode.Stretch)
-        self.channel_tableWidget.verticalHeader().setVisible(False)
-        self.channel_tableWidget.setTabKeyNavigation(True)
-        self.channel_tableWidget.setColumnCount(2)
-        self.channel_tableWidget.setRowCount(0)
-        self.channel_tableWidget.setHorizontalHeaderLabels(
-            ["Channel", "Exposure Time (ms)"]
-        )
-        group_layout.addWidget(self.channel_tableWidget, 0, 0)
-
-        # buttons
-        wdg = QWidget()
-        layout = QVBoxLayout()
-        layout.setSpacing(10)
-        layout.setContentsMargins(0, 0, 0, 0)
-        wdg.setLayout(layout)
-
-        btn_sizepolicy = QSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Fixed)
-        min_size = 100
-        self.add_ch_button = QPushButton(text="Add")
-        self.add_ch_button.setMinimumWidth(min_size)
-        self.add_ch_button.setSizePolicy(btn_sizepolicy)
-        self.remove_ch_button = QPushButton(text="Remove")
-        self.remove_ch_button.setMinimumWidth(min_size)
-        self.remove_ch_button.setSizePolicy(btn_sizepolicy)
-        self.clear_ch_button = QPushButton(text="Clear")
-        self.clear_ch_button.setMinimumWidth(min_size)
-        self.clear_ch_button.setSizePolicy(btn_sizepolicy)
-
-        self.add_ch_button.clicked.connect(self._add_channel)
-        self.remove_ch_button.clicked.connect(self._remove_channel)
-        self.clear_ch_button.clicked.connect(self._clear_channel)
-
-        spacer = QSpacerItem(
-            10, 0, QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Expanding
-        )
-
-        layout.addWidget(self.add_ch_button)
-        layout.addWidget(self.remove_ch_button)
-        layout.addWidget(self.clear_ch_button)
-        layout.addItem(spacer)
-
-        group_layout.addWidget(wdg, 0, 1)
-
-    def _add_channel(self) -> bool:
-        """Add, remove or clear channel table.  Return True if anyting was changed."""
-        if len(self._mmc.getLoadedDevices()) <= 1:
-            return False
-
-        channel_group = self._mmc.getChannelGroup()
-        if not channel_group:
-            return False
-
-        idx = self.channel_tableWidget.rowCount()
-        self.channel_tableWidget.insertRow(idx)
-
-        # create a combo_box for channels in the table
-        channel_comboBox = QComboBox(self)
-        channel_exp_spinBox = QSpinBox(self)
-        channel_exp_spinBox.setRange(0, 10000)
-        channel_exp_spinBox.setValue(100)
-        channel_exp_spinBox.valueChanged.connect(self._on_exp_changed)
-
-        if channel_group := self._mmc.getChannelGroup():
-            channel_list = list(self._mmc.getAvailableConfigs(channel_group))
-            channel_comboBox.addItems(channel_list)
-
-        self.channel_tableWidget.setCellWidget(idx, 0, channel_comboBox)
-        self.channel_tableWidget.setCellWidget(idx, 1, channel_exp_spinBox)
-
-        self.valueUpdated.emit()
-
-        return True
-
-    def _on_exp_changed(self) -> None:
-        self.valueUpdated.emit()
-
-    def _remove_channel(self) -> None:
-        # remove selected position
-        rows = {r.row() for r in self.channel_tableWidget.selectedIndexes()}
-        for idx in sorted(rows, reverse=True):
-            self.channel_tableWidget.removeRow(idx)
-
-        self.valueUpdated.emit()
-
-    def _clear_channel(self) -> None:
-        # clear all positions
-        self.channel_tableWidget.clearContents()
-        self.channel_tableWidget.setRowCount(0)
-
-        self.valueUpdated.emit()
-
-
 class _MDATimeWidget(QGroupBox):
 
     valueUpdated = Signal()
