+++ src/pymmcore_widgets/_sample_explorer_widget/_sample_explorer_widget.py
@@ -73,7 +73,7 @@ class SampleExplorerWidget(SampleExplorerGui):
         self.pos_gp = self.stage_pos_groupbox
 
         # connect valueUpdated signal
-        self.ch_gb.valueUpdated.connect(self._update_total_time)
+        self.ch_gb.valueChanged.connect(self._update_total_time)
         self.z_gp.valueChanged.connect(self._update_total_time)
         self.tm_gp.valueUpdated.connect(self._update_total_time)
 
@@ -429,16 +429,7 @@ class SampleExplorerWidget(SampleExplorerGui):
         -------
         useq.MDASequence
         """
-        table = self.ch_gb.channel_tableWidget
-
-        channels: list[dict] = [
-            {
-                "config": table.cellWidget(c, 0).currentText(),
-                "group": self._mmc.getChannelGroup() or "Channel",
-                "exposure": table.cellWidget(c, 1).value(),
-            }
-            for c in range(table.rowCount())
-        ]
+        channels = self.channel_groupbox.value()
 
         z_plan = self.z_gp.value() if self.z_gp.isChecked() else None
